+++ core/api/typescript/src/kurtosis_core_rpc_api_bindings/api_container_service_grpc_web_pb.js
@@ -1170,5 +1170,66 @@ proto.api_container_api.ApiContainerServicePromiseClient.prototype.storeFilesArt
 };
 
 
+/**
+ * @const
+ * @type {!grpc.web.MethodDescriptor<
+ *   !proto.api_container_api.RenderTemplatesToFilesArtifactArgs,
+ *   !proto.api_container_api.RenderTemplatesToFilesArtifactResponse>}
+ */
+const methodDescriptor_ApiContainerService_RenderTemplatesToFilesArtifact = new grpc.web.MethodDescriptor(
+  '/api_container_api.ApiContainerService/RenderTemplatesToFilesArtifact',
+  grpc.web.MethodType.UNARY,
+  proto.api_container_api.RenderTemplatesToFilesArtifactArgs,
+  proto.api_container_api.RenderTemplatesToFilesArtifactResponse,
+  /**
+   * @param {!proto.api_container_api.RenderTemplatesToFilesArtifactArgs} request
+   * @return {!Uint8Array}
+   */
+  function(request) {
+    return request.serializeBinary();
+  },
+  proto.api_container_api.RenderTemplatesToFilesArtifactResponse.deserializeBinary
+);
+
+
+/**
+ * @param {!proto.api_container_api.RenderTemplatesToFilesArtifactArgs} request The
+ *     request proto
+ * @param {?Object<string, string>} metadata User defined
+ *     call metadata
+ * @param {function(?grpc.web.RpcError, ?proto.api_container_api.RenderTemplatesToFilesArtifactResponse)}
+ *     callback The callback function(error, response)
+ * @return {!grpc.web.ClientReadableStream<!proto.api_container_api.RenderTemplatesToFilesArtifactResponse>|undefined}
+ *     The XHR Node Readable Stream
+ */
+proto.api_container_api.ApiContainerServiceClient.prototype.renderTemplatesToFilesArtifact =
+    function(request, metadata, callback) {
+  return this.client_.rpcCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/RenderTemplatesToFilesArtifact',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_RenderTemplatesToFilesArtifact,
+      callback);
+};
+
+
+/**
+ * @param {!proto.api_container_api.RenderTemplatesToFilesArtifactArgs} request The
+ *     request proto
+ * @param {?Object<string, string>=} metadata User defined
+ *     call metadata
+ * @return {!Promise<!proto.api_container_api.RenderTemplatesToFilesArtifactResponse>}
+ *     Promise that resolves to the response
+ */
+proto.api_container_api.ApiContainerServicePromiseClient.prototype.renderTemplatesToFilesArtifact =
+    function(request, metadata) {
+  return this.client_.unaryCall(this.hostname_ +
+      '/api_container_api.ApiContainerService/RenderTemplatesToFilesArtifact',
+      request,
+      metadata || {},
+      methodDescriptor_ApiContainerService_RenderTemplatesToFilesArtifact);
+};
+
+
 module.exports = proto.api_container_api;