+++ core/docs/changelog.md
@@ -1,5 +1,9 @@
 # TBD
 
+### Features
+* Container engine server exposes a `RenderTemplatesToFilesArtifact` endpoint that takes in templates & data and returns a file artifact.
+* Added `RemoveFile` to both `FileCache` and `FilesArtifactStore` to support removing files from the cache and store.
+
 # 1.58.1
 ### Changes
 * Replaced the `matchall` filter type with the `basic` filter type in network partitioning commands due to some OSs (e.g.: Apache Linux distribution) not having the `cls_matchall` Linux kernel's module